server/diagnostics: move dep_test.go to the right package, fix dep
Release note: None
