+++ pkg/server/BUILD.bazel
@@ -93,7 +93,7 @@ go_library(
         "//pkg/security",
         "//pkg/server/debug",
         "//pkg/server/diagnostics",
-        "//pkg/server/diagnostics/diagnosticspb",
+        "//pkg/server/diagnostics/diagnosticspb:diagnosticspb_go_proto",
         "//pkg/server/goroutinedumper",
         "//pkg/server/heapprofiler",
         "//pkg/server/serverpb",
@@ -297,7 +297,7 @@ go_test(
         "//pkg/security/securitytest",
         "//pkg/server/debug",
         "//pkg/server/diagnostics",
-        "//pkg/server/diagnostics/diagnosticspb",
+        "//pkg/server/diagnostics/diagnosticspb:diagnosticspb_go_proto",
         "//pkg/server/serverpb",
         "//pkg/server/status/statuspb:statuspb_go_proto",
         "//pkg/server/telemetry",
