+++ pkg/server/diagnostics/BUILD.bazel
@@ -17,8 +17,8 @@ go_library(
         "//pkg/kv",
         "//pkg/roachpb",
         "//pkg/security",
-        "//pkg/server/diagnostics/diagnosticspb",
-        "//pkg/server/status",
+        "//pkg/server/diagnostics/diagnosticspb:diagnosticspb_go_proto",
+        "//pkg/server/status/statuspb:statuspb_go_proto",
         "//pkg/server/telemetry",
         "//pkg/settings",
         "//pkg/settings/cluster",
@@ -51,12 +51,13 @@ go_test(
     name = "diagnostics_test",
     size = "medium",
     srcs = [
+        "dep_test.go",
         "main_test.go",
         "reporter_test.go",
         "update_checker_test.go",
     ],
+    embed = [":diagnostics"],
     deps = [
-        ":diagnostics",
         "//pkg/base",
         "//pkg/build",
         "//pkg/ccl/kvccl/kvtenantccl",
@@ -67,12 +68,13 @@ go_test(
         "//pkg/security",
         "//pkg/security/securitytest",
         "//pkg/server",
-        "//pkg/server/diagnostics/diagnosticspb",
+        "//pkg/server/diagnostics/diagnosticspb:diagnosticspb_go_proto",
         "//pkg/server/telemetry",
         "//pkg/settings/cluster",
         "//pkg/sql",
         "//pkg/sql/catalog/lease",
         "//pkg/testutils",
+        "//pkg/testutils/buildutil",
         "//pkg/testutils/diagutils",
         "//pkg/testutils/serverutils",
         "//pkg/testutils/skip",
