+++ pkg/server/diagnostics/diagnosticspb/BUILD.bazel
@@ -1,5 +1,4 @@
 load("@rules_proto//proto:defs.bzl", "proto_library")
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 load("@io_bazel_rules_go//proto:def.bzl", "go_proto_library")
 
 proto_library(
@@ -29,23 +28,5 @@ go_proto_library(
         "//pkg/roachpb",
         "//pkg/sql/catalog/descpb",
         "@com_github_gogo_protobuf//gogoproto",
-    ],
-)
-
-go_library(
-    name = "diagnosticspb",
-    embed = [":diagnosticspb_go_proto"],
-    importpath = "github.com/cockroachdb/cockroach/pkg/server/diagnostics/diagnosticspb",
-    visibility = ["//visibility:public"],
-)
-
-go_test(
-    name = "diagnosticspb_test",
-    size = "small",
-    srcs = ["dep_test.go"],
-    embed = [":diagnosticspb"],
-    deps = [
-        "//pkg/testutils/buildutil",
-        "//pkg/util/leaktest",
     ],
 )
