+++ pkg/server/serverpb/BUILD.bazel
@@ -58,7 +58,7 @@ go_proto_library(
         "//pkg/kv/kvserver/kvserverpb",
         "//pkg/kv/kvserver/liveness/livenesspb",
         "//pkg/roachpb",
-        "//pkg/server/diagnostics/diagnosticspb",
+        "//pkg/server/diagnostics/diagnosticspb:diagnosticspb_go_proto",
         "//pkg/server/status/statuspb:statuspb_go_proto",
         "//pkg/sql/catalog/descpb",  # keep
         "//pkg/sql/contentionpb",
