+++ pkg/testutils/diagutils/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "github.com/cockroachdb/cockroach/pkg/testutils/diagutils",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/server/diagnostics/diagnosticspb",
+        "//pkg/server/diagnostics/diagnosticspb:diagnosticspb_go_proto",
         "//pkg/util/protoutil",
         "//pkg/util/syncutil",
     ],
